Feat: Odata Batch Request Headers validation
If the set of request headers of a batch request are valid the service MUST return a 200 OK HTTP response code to indicate that the batch request was accepted for processing, but the processing is yet to be completed. The individual requests within the body of the batch request may subsequently fail or be malformed; however, this enables batch implementations to stream the results.
If the service receives a batch request with an invalid set of headers it MUST return a 4xx response code and perform no further processing of the batch request.